review requirements
reviewer: As the previous reviewer mentioned. For each displayed project, the description of each of these projects is on top of their image with a dark opacity with white color text. when the project is hovered on with the mouse the description and dark opacity disappear and a See Project button appears. Kindly fix the issue, you can achieve this using CSS by changing the display's value on :hover of these elements.

- implemented this feature to all the cards